+++ CHANGELOG.md
@@ -49,6 +49,7 @@ All notable changes to this project will be documented in this file.
 - \[[e6ab5ca9](https://github.com/silveiralexf/.dotfiles/commit/e6ab5ca9c93552cb0a41e42d020a55f78a2a2460)\] update cursor workflow, agents, rules, commands and plugin config
 - \[[fecfc007](https://github.com/silveiralexf/.dotfiles/commit/fecfc007469dcb49d06d9d5b766fc6e1fe283abd)\] update nvim lazy lock, lazyvim and plugin config
 - \[[8334f3d8](https://github.com/silveiralexf/.dotfiles/commit/8334f3d88e163c3425a649e8c3390bf9ce6243eb)\] add Lua tests (task test:lua) to pre-commit hooks
+- \[[528a2b14](https://github.com/silveiralexf/.dotfiles/commit/528a2b148fc2553643825d40b76d87dcf365c396)\] improve Lua test output (TAP) and drop reference name from docs
 
 
 ## 2026-01-09